delete krypto only if it's defined

+++ js/src/engine/KryptnosticEngineProvider.coffee
@@ -26,7 +26,8 @@ define 'kryptnostic.kryptnostic-engine-provider', [
       throw new Error('KryptnosticEngine has not been initialized')
 
     @destroy: ->
-      _engine.krypto.delete()
+      if _engine.krypto and _engine.krypto.delete
+        _engine.krypto.delete()
       _engine = null
       return
